change how comment is parsed
