+++ .eleventy.js
@@ -2,6 +2,7 @@
 const cleanCSS = require('clean-css');
 const eleventyNavigationPlugin = require('@11ty/eleventy-navigation');
 const { escapeHtml } = require('markdown-it/lib/common/utils');
+const { parseConfigurationValue } = require('./tools/eleventy-utils');
 
 module.exports = function(eleventyConfig) {
 	eleventyConfig.addPassthroughCopy('img');
@@ -66,29 +67,7 @@ module.exports = function(eleventyConfig) {
 
 	markdownIt.renderer.rules.fence = (tokens, idx) => {
 		const content = tokens[idx].content;
-		if (content.includes('<!-- docs: live demo') || content.includes('<!-- docs: demo -->') || content.includes('<!-- docs: code demo -->')) {
-			if (content.includes('<!-- docs: live demo')) {
-				return `
-					<d2l-component-catalog-demo-snippet interactive resizable demo-snippet="${escapeHtml(content)}">
-					</d2l-component-catalog-demo-snippet>
-				`;
-			} else if (content.includes('<!-- docs: code demo -->')) {
-
-				return `
-					<d2l-component-catalog-demo-snippet resizable demo-snippet="${escapeHtml(content)}">
-					</d2l-component-catalog-demo-snippet>
-				`;
-
-			} else {
-				return `
-					<d2l-component-catalog-demo-snippet resizable hide-code demo-snippet="${escapeHtml(content)}">
-					</d2l-component-catalog-demo-snippet>
-				`;
-			}
-		} else {
-			// Code only snippets
-			return `<d2l-component-catalog-demo-snippet code-only demo-snippet="${escapeHtml(content)}"></d2l-component-catalog-demo-snippet>`;
-		}
+		return `${escapeHtml(content)}</d2l-component-catalog-demo-snippet>`;
 	};
 
 	markdownIt.renderer.rules.table_open = () => {
@@ -118,7 +97,20 @@ module.exports = function(eleventyConfig) {
 	const defaultHtmlRule = markdownIt.renderer.rules.html_block;
 	markdownIt.renderer.rules.html_block = (tokens, idx, options, env, slf) => {
 		const content = tokens[idx].content;
-		if (content.includes('<!-- docs: start hidden content -->'))
+		if (content.includes('<!-- docs: live demo')) {
+			const tag = parseConfigurationValue('name', content);
+			if (!tag) return '<d2l-component-catalog-demo-snippet resizable hide-code>';
+			const size = parseConfigurationValue('size', content);
+			const defaults = parseConfigurationValue('defaults', content, true);
+			let openingTag = `<d2l-component-catalog-demo-snippet resizable interactive tag-name="${tag}" `;
+			if (size) openingTag += ` size="${size}" `;
+			if (defaults) openingTag += ` defaults='${defaults}'`;
+			return `${openingTag}>`;
+		} else if (content.includes('<!-- docs: code demo -->'))
+			return '<d2l-component-catalog-demo-snippet resizable>';
+		else if (content.includes('<!-- docs: demo -->'))
+			return '<d2l-component-catalog-demo-snippet resizable hide-code>';
+		else if (content.includes('<!-- docs: start hidden content -->'))
 			return '<div style="display: none;">';
 		else if (content.includes('<!-- docs: end'))
 			return '</div>';
